feat: Setup redux-observable
- Change directory reducer to redux
- move redux store into redux/inde
