+++ client/src/components/StatusBar.tsx
@@ -4,7 +4,7 @@ import { jsx, css } from "@emotion/core";
 import { Link } from "react-router-dom";
 import useCurrentUser from "../hooks/useCurrentUser";
 import { useDispatch } from "react-redux";
-import { startEdit } from "../reducers/editorReducer";
+import { startEdit } from "../redux/editorReducer";
 
 const StatusBarCss = css`
   display: flex;
