+++ client/src/components/main-page/__test__/ZettelViewerPanel.test.tsx
@@ -4,7 +4,7 @@ import { useDispatch } from "react-redux";
 
 import { initZettels } from "../../../__mocks__/react-redux";
 import ZettelViewerPanel from "../ZettelViewerPanel";
-import { setViewer } from "../../../reducers/editorReducer";
+import { setViewer } from "../../../redux/editorReducer";
 
 jest.mock("react-redux");
