+++ client/src/redux/epics/index.ts
@@ -1,0 +1,3 @@
+import { combineEpics } from "redux-observable";
+const rootEpic = combineEpics();
+export default rootEpic;
